requirement.txt
opencv-python
matplotlib
numpy
Pillow
torch
torchvision
tqdm
pyyaml
requests
pandas
seaborn
mysql-connector-python
paho-mqtt